Refactor: Separate general SPDX license tools to those for completion
